requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiounittest==1.3.0
applicationinsights==0.11.9
asgiref==3.6.0
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.4.1
attrs==22.2.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
azure-core==1.26.3
azure-cosmos==3.2.0
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
Babel==2.9.1
botbuilder-ai==4.14.3
botbuilder-applicationinsights==4.14.3
botbuilder-azure==4.14.3
botbuilder-core==4.14.3
botbuilder-dialogs==4.14.3
botbuilder-integration-applicationinsights-aiohttp==4.14.3
botbuilder-schema==4.14.3
botbuilder-testing==4.14.3
botframework-connector==4.14.3
botframework-streaming==4.14.3
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
datatypes-date-time==1.0.0a2
datedelta==1.4
Django==3.2.16
djangorestframework==3.10.3
emoji==1.7.0
Flask==1.1.1
google-api-core==2.11.0
google-auth==2.16.2
googleapis-common-protos==1.58.0
grapheme==0.6.0
idna==2.10
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==1.4.2
MarkupSafe==2.1.2
msal==1.20.0
msal-extensions==1.0.0
msrest==0.6.21
msrestazure==0.6.4
multidict==6.0.4
multipledispatch==0.6.0
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opencensus-ext-azure==1.1.8
packaging==23.0
pluggy==1.0.0
portalocker==2.7.0
protobuf==4.22.0
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.25.11
Werkzeug==2.2.2
yarl==1.8.2
